feat: refractor grpc
