+++ CMakeLists.txt
@@ -73,6 +73,12 @@ add_library(${PROJECT_NAME} SHARED
   "src/${PROJECT_NAME}/walking/node/walking_node.cpp"
   "src/${PROJECT_NAME}/walking/process/kinematic.cpp"
   "src/${PROJECT_NAME}/config/grpc/config.cpp"
+  "src/${PROJECT_NAME}/config/grpc/call_data_base.cpp"
+  "src/${PROJECT_NAME}/config/grpc/call_data.cpp"
+  "src/${PROJECT_NAME}/config/grpc/call_data_get_config.cpp"
+  "src/${PROJECT_NAME}/config/grpc/call_data_set_config.cpp"
+  "src/${PROJECT_NAME}/config/grpc/call_data_save_config.cpp"
+  "src/${PROJECT_NAME}/config/grpc/call_data_publish_config.cpp"
   "${aruku_proto_srcs}"
   "${aruku_proto_hdrs}"
   "${aruku_grpc_srcs}"
