+++ include/aruku/config/grpc/config.hpp
@@ -43,6 +43,8 @@
 #include "tachimawari/joint/model/joint.hpp"
 #include "aruku_interfaces/msg/set_config.hpp"
 #include "aruku_interfaces/msg/set_walking.hpp"
+#include "aruku/config/grpc/call_data_base.hpp"
+#include "aruku/config/grpc/call_data.hpp"
 
 using aruku_interfaces::proto::Config;
 
@@ -61,106 +63,6 @@ public:
 private:
   std::string path;  
   static void SignIntHandler(int signum);
-
-  class CallDataBase
-  {
-  public:
-    CallDataBase();
-
-    virtual void Proceed() = 0;
-
-  protected:
-    virtual void WaitForRequest() = 0;
-    virtual void HandleRequest() = 0;
-  };
-
-  template <class ConfigRequest, class ConfigReply>
-  class CallData : CallDataBase
-  {
-  public:
-    CallData(
-      aruku_interfaces::proto::Config::AsyncService * service, grpc::ServerCompletionQueue * cq,
-      const std::string path);
-
-    virtual void Proceed() override;
-
-  protected:
-    virtual void AddNextToCompletionQueue() = 0;
-
-    enum CallStatus { CREATE, PROCESS, FINISH };
-
-    CallStatus status_;  // The current serving state.
-
-    aruku_interfaces::proto::Config::AsyncService * service_;
-
-    const std::string path_;
-
-    grpc::ServerCompletionQueue * cq_;
-    grpc::ServerContext ctx_;
-    ConfigRequest request_;
-    ConfigReply reply_;
-    grpc::ServerAsyncResponseWriter<ConfigReply> responder_;
-  };
-
-  class CallDataGetConfig
-  : CallData<aruku_interfaces::proto::Empty, aruku_interfaces::proto::ConfigWalking>
-  {
-  public:
-    CallDataGetConfig(
-      aruku_interfaces::proto::Config::AsyncService * service, grpc::ServerCompletionQueue * cq,
-      const std::string path);
-
-  protected:
-    virtual void AddNextToCompletionQueue() override;
-    virtual void WaitForRequest() override;
-    virtual void HandleRequest() override;
-  };
-
-  class CallDataSaveConfig
-  : CallData<aruku_interfaces::proto::ConfigWalking, aruku_interfaces::proto::Empty>
-  {
-  public:
-    CallDataSaveConfig(
-      aruku_interfaces::proto::Config::AsyncService * service, grpc::ServerCompletionQueue * cq,
-      const std::string path);
-
-  protected:
-    virtual void AddNextToCompletionQueue() override;
-    virtual void WaitForRequest() override;
-    virtual void HandleRequest() override;
-  };
-
-  class CallDataPublishConfig
-  : CallData<aruku_interfaces::proto::ConfigWalking, aruku_interfaces::proto::Empty>
-  {
-  public:
-    CallDataPublishConfig(
-      aruku_interfaces::proto::Config::AsyncService * service, grpc::ServerCompletionQueue * cq,
-      const std::string path, rclcpp::Node::SharedPtr node);
-
-  protected:
-    rclcpp::Node::SharedPtr node_;
-    rclcpp::Publisher<aruku_interfaces::msg::SetConfig>::SharedPtr set_config_publisher_;
-    virtual void AddNextToCompletionQueue() override;
-    virtual void WaitForRequest() override;
-    virtual void HandleRequest() override;    
-  };
-
-  class CallDataSetConfig
-  : CallData<aruku_interfaces::proto::SetWalking, aruku_interfaces::proto::Empty>
-  {
-  public:
-    CallDataSetConfig(
-      aruku_interfaces::proto::Config::AsyncService * service, grpc::ServerCompletionQueue * cq,
-      const std::string path, rclcpp::Node::SharedPtr node);
-
-  protected:    
-    virtual void AddNextToCompletionQueue() override;
-    virtual void WaitForRequest() override;
-    virtual void HandleRequest() override;    
-    rclcpp::Node::SharedPtr node_;
-    rclcpp::Publisher<aruku_interfaces::msg::SetWalking>::SharedPtr set_config_publisher_;
-  };
   
   static inline std::unique_ptr<grpc::ServerCompletionQueue> cq_;
   static inline std::unique_ptr<grpc::Server> server_;
